Update decredPoS_script.md
Just a few recommendations (in the final hour) 

ticketholder > stakeholder in anticipation for broader ticket ownership (ie: ticket splitting) as well as the reiteration of 'stake'  and it's association with holding tickets. 

"current data" - current isn't really a word to stand the tests of time - this current data figure loosely thrown around is based on the voting time based on the target pool size ~28-31 days.   

removal of some of the heavy terms ("there are a time and a place"), simply not appropriate for the first educational video in my honest opinion. 

re-structuring of the introduction for better flow and emphasis of hybridised system with less "bitcoin-bashing" as there are far worse offenders out there.

ending - reworked - should result in a snappier delivery to wrap things up.